Replaces LinkList with FooterLink
