+++ components/ui/footer/FooterSubsection.vue
@@ -3,19 +3,21 @@
     <h2 class="footer-title">
       {{ title }}
     </h2>
-    <LinkList :elements="elements" />
+    <FooterLink
+      v-for="element in elements"
+      :key="element.label"
+      v-bind="element"
+    />
   </div>
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
 import { Component, Prop } from 'vue-property-decorator'
-import LinkList from '~/components/ui/footer/LinkList.vue'
+import FooterLink from '~/components/ui/footer/FooterLink.vue'
 
 @Component({
-  components: {
-    LinkList
-  }
+  components: { FooterLink }
 })
 export default class extends Vue {
   @Prop(String) title
@@ -29,6 +31,7 @@ export default class extends Vue {
 .footer-title {
   @include type-style('productive-heading-03');
   margin-top: 2rem;
+  margin-bottom: 1.5rem;
   padding-bottom: 0.5rem;
   border-bottom: 1px solid $ui-03;
 }
