+++ constants/menuLinks.ts
@@ -164,7 +164,7 @@ const ORDERED_COMMUNITY_SUB_LINKS: Array<NavLink> = [
 const QISKIT_EDUCATION: LabeledLinks = {
   textbook: {
     label: 'Textbook',
-    url: '/textbook',
+    url: 'https://qiskit.org/textbook/preface.html',
     linkType: 'aLink'
   },
   youtube: {
